requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
geopandas>=0.14.2
pydeck>=0.8.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
SQLAlchemy==2.0.29
plotly==5.18.0
numpy==1.26.4 